+++ src/openassetio-python/cmodule/src/managerApi/ManagerInterfaceBinding.cpp
@@ -46,8 +46,7 @@ struct PyManagerInterface : ManagerInterface {
   }
 
   void initialize(InfoDictionary managerSettings, const HostSessionPtr& hostSession) override {
-    PYBIND11_OVERRIDE_PURE(void, ManagerInterface, initialize, std::move(managerSettings),
-                           hostSession);
+    PYBIND11_OVERRIDE(void, ManagerInterface, initialize, std::move(managerSettings), hostSession);
   }
 
   void flushCaches(const HostSessionPtr& hostSession) override {
@@ -58,8 +57,8 @@ struct PyManagerInterface : ManagerInterface {
                                                     access::PolicyAccess policyAccess,
                                                     const ContextConstPtr& context,
                                                     const HostSessionPtr& hostSession) override {
-    PYBIND11_OVERRIDE_PURE(trait::TraitsDatas, ManagerInterface, managementPolicy, traitSets,
-                           policyAccess, context, hostSession);
+    PYBIND11_OVERRIDE(trait::TraitsDatas, ManagerInterface, managementPolicy, traitSets,
+                      policyAccess, context, hostSession);
   }
 
   ManagerStateBasePtr createState(const HostSessionPtr& hostSession) override {
@@ -85,16 +84,15 @@ struct PyManagerInterface : ManagerInterface {
 
   [[nodiscard]] bool isEntityReferenceString(const Str& someString,
                                              const HostSessionPtr& hostSession) override {
-    PYBIND11_OVERRIDE_PURE(bool, ManagerInterface, isEntityReferenceString, someString,
-                           hostSession);
+    PYBIND11_OVERRIDE(bool, ManagerInterface, isEntityReferenceString, someString, hostSession);
   }
 
   void entityExists(const EntityReferences& entityReferences, const ContextConstPtr& context,
                     const HostSessionPtr& hostSession,
                     const ExistsSuccessCallback& successCallback,
                     const BatchElementErrorCallback& errorCallback) override {
-    PYBIND11_OVERRIDE_PURE(void, ManagerInterface, entityExists, entityReferences, context,
-                           hostSession, successCallback, errorCallback);
+    PYBIND11_OVERRIDE(void, ManagerInterface, entityExists, entityReferences, context, hostSession,
+                      successCallback, errorCallback);
   }
 
   [[nodiscard]] bool hasCapability(ManagerInterface::Capability capability) override {
@@ -110,8 +108,8 @@ struct PyManagerInterface : ManagerInterface {
                const access::ResolveAccess resolveAccess, const ContextConstPtr& context,
                const HostSessionPtr& hostSession, const ResolveSuccessCallback& successCallback,
                const BatchElementErrorCallback& errorCallback) override {
-    PYBIND11_OVERRIDE_PURE(void, ManagerInterface, resolve, entityReferences, traitSet,
-                           resolveAccess, context, hostSession, successCallback, errorCallback);
+    PYBIND11_OVERRIDE(void, ManagerInterface, resolve, entityReferences, traitSet, resolveAccess,
+                      context, hostSession, successCallback, errorCallback);
   }
 
   void defaultEntityReference(const trait::TraitSets& traitSets,
@@ -158,16 +156,17 @@ struct PyManagerInterface : ManagerInterface {
                  const HostSessionPtr& hostSession,
                  const PreflightSuccessCallback& successCallback,
                  const BatchElementErrorCallback& errorCallback) override {
-    PYBIND11_OVERRIDE_PURE(void, ManagerInterface, preflight, entityReferences, traitsHints,
-                           publishingAccess, context, hostSession, successCallback, errorCallback);
+    PYBIND11_OVERRIDE(void, ManagerInterface, preflight, entityReferences, traitsHints,
+                      publishingAccess, context, hostSession, successCallback, errorCallback);
   }
 
   void register_(const EntityReferences& entityReferences, const trait::TraitsDatas& traitsDatas,
                  const access::PublishingAccess publishingAccess, const ContextConstPtr& context,
                  const HostSessionPtr& hostSession, const RegisterSuccessCallback& successCallback,
                  const BatchElementErrorCallback& errorCallback) override {
-    PYBIND11_OVERRIDE_PURE(void, ManagerInterface, register, entityReferences, traitsDatas,
-                           publishingAccess, context, hostSession, successCallback, errorCallback);
+    PYBIND11_OVERRIDE_NAME(void, ManagerInterface, "register", register_, entityReferences,
+                           traitsDatas, publishingAccess, context, hostSession, successCallback,
+                           errorCallback);
   }
 
   // Hoist protected members